refactor: 💡 SwaggerConfig 생성 후 관련 설정 이동
