+++ src/main/java/com/postsquad/scoup/ScoupApplication.java
@@ -1,42 +1,14 @@
 package com.postsquad.scoup;
 
-import com.google.common.base.Predicates;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import springfox.documentation.builders.ApiInfoBuilder;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.service.ApiInfo;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
-@EnableSwagger2
 @EnableJpaAuditing
 @SpringBootApplication
 public class ScoupApplication {
 
     public static void main(String[] args) {
         SpringApplication.run(ScoupApplication.class, args);
-    }
-
-    @Bean
-    public Docket api() {
-        return new Docket(DocumentationType.SWAGGER_2)
-                .groupName("Scoup")
-                .apiInfo(apiInfo())
-                .select()
-                .paths(Predicates.not(PathSelectors.regex("/error")))
-                .paths(PathSelectors.ant("/**"))
-                .build();
-    }
-
-    private ApiInfo apiInfo() {
-        return new ApiInfoBuilder()
-                .title("Scoup API Documentation")
-                .version("1.0")
-                .description("Scoup의 API 문서입니다.")
-                .build();
     }
 }
